requirements.txt
python-decouple==3.8
python-dotenv==1.0.0
colorama==0.4.6
pytest==7.4.3
black==24.2.0
flake8==7.0.0
mypy==1.8.0
pytest-cov==4.1.0 